Conflicts:
	src/libxsmm_blkgemm.c

+++ src/libxsmm_ext_trans.c
@@ -50,10 +50,10 @@ LIBXSMM_API_DEFINITION int libxsmm_matcopy_omp(void* out, const void* in, unsign
 
   assert(typesize <= 255);
   if (0 != out && out != in && 0 < typesize && 0 < m && 0 < n && m <= ldi && n <= ldo) {
-    const unsigned int size = 1U * m * n;
+    const unsigned int size = (unsigned int)(1U * m * n);
     if (size > (LIBXSMM_TRANS_THRESHOLD)) { /* consider problem-size (threshold) */
       const int tindex = (4 < typesize ? 0 : 1), index = LIBXSMM_MIN(LIBXSMM_SQRT2(size) >> 10, 7);
-      const unsigned int uldi = ldi, uldo = ldo;
+      const unsigned int uldi = (unsigned int)ldi, uldo = (unsigned int)ldo;
       libxsmm_matcopy_descriptor descriptor = { 0 };
       libxsmm_xmatcopyfunction xmatcopy = 0;
       LIBXSMM_INIT
@@ -62,7 +62,7 @@ LIBXSMM_API_DEFINITION int libxsmm_matcopy_omp(void* out, const void* in, unsign
       descriptor.prefetch = (unsigned char)((0 == prefetch || 0 == *prefetch) ? 0 : 1);
       if (0 != (1 & libxsmm_trans_jit)) { /* JIT'ted matcopy permitted? */
         descriptor.flags = (unsigned char)(0 != in ? 0 : LIBXSMM_MATCOPY_FLAG_ZERO_SOURCE);
-        descriptor.ldi = ldi; descriptor.ldo = ldo; descriptor.unroll_level = 2;
+        descriptor.ldi = (unsigned int)ldi; descriptor.ldo = (unsigned int)ldo; descriptor.unroll_level = 2;
         descriptor.typesize = (unsigned char)typesize;
         xmatcopy = libxsmm_xmatcopydispatch(&descriptor);
       }
@@ -146,16 +146,16 @@ LIBXSMM_API_DEFINITION int libxsmm_otrans_omp(void* out, const void* in, unsigne
   if (0 != out && 0 != in && 0 < typesize && 0 < m && 0 < n && m <= ldi && n <= ldo) {
     LIBXSMM_INIT
     if (out != in) {
-      const unsigned int size = 1U * m * n;
+      const unsigned int size = (unsigned int)(1U * m * n);
       if (size > (LIBXSMM_TRANS_THRESHOLD)) { /* consider problem-size (threshold) */
         const int tindex = (4 < typesize ? 0 : 1), index = LIBXSMM_MIN(LIBXSMM_SQRT2(size) >> 10, 7);
-        const unsigned int uldi = ldi, uldo = ldo;
+        const unsigned int uldi = (unsigned int)ldi, uldo = (unsigned int)ldo;
         libxsmm_transpose_descriptor descriptor = { 0 };
         libxsmm_xtransfunction xtrans = 0;
         descriptor.m = LIBXSMM_MIN((unsigned int)m, libxsmm_trans_tile[tindex][0/*M*/][index]);
         descriptor.n = LIBXSMM_MIN((unsigned int)n, libxsmm_trans_tile[tindex][1/*N*/][index]);
         if (0 != (2 & libxsmm_trans_jit)) { /* JIT'ted transpose */
-          descriptor.typesize = (unsigned char)typesize; descriptor.ldo = ldo;
+          descriptor.typesize = (unsigned char)typesize; descriptor.ldo = (unsigned int)ldo;
           descriptor.m = LIBXSMM_MIN(descriptor.m, LIBXSMM_MAX_M);
           descriptor.n = LIBXSMM_MIN(descriptor.n, LIBXSMM_MAX_N);
           xtrans = libxsmm_xtransdispatch(&descriptor);
